+++ composer.json
@@ -22,7 +22,7 @@
     },
     "autoload": {
         "psr-0": {
-            "React\\Nntp": "src/"
+            "Rvdv\\React\\Nntp": "src/"
         }
     },
     "config": {
