+++ phpunit.xml.dist
@@ -13,7 +13,7 @@
 >
     <testsuites>
         <testsuite name="React/NNTP Test Suite">
-            <directory>./tests/React/</directory>
+            <directory>./tests/Rvdv/React/</directory>
         </testsuite>
     </testsuites>
