--- src/React/Nntp/Exception/ConnectionException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace React\Nntp\Exception;
-
-class ConnectionException extends \ErrorException
-{
-}
