--- src/React/Nntp/Response/MultilineResponseInterface.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace React\Nntp\Response;
-
-/**
- * MultilineResponseInterface
- *
- * @author Robin van der Vleuten <robinvdvleuten@gmail.com>
- */
-interface MultilineResponseInterface extends ResponseInterface
-{
-    public function getLines();
-}
