+++ src/Rvdv/React/Nntp/Exception/ConnectionException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Rvdv\React\Nntp\Exception;
+
+class ConnectionException extends \ErrorException
+{
+}
