+++ src/Rvdv/React/Nntp/Response/MultilineResponseInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Rvdv\React\Nntp\Response;
+
+/**
+ * MultilineResponseInterface
+ *
+ * @author Robin van der Vleuten <robinvdvleuten@gmail.com>
+ */
+interface MultilineResponseInterface extends ResponseInterface
+{
+    public function getLines();
+}
