--- tests/React/Tests/Nntp/Stub/CallableStub.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace React\Tests\Nntp\Stub;
-
-class CallableStub
-{
-    public function __invoke()
-    {
-    }
-}
