--- tests/React/Tests/Nntp/TestCase.php
@@ -1,41 +1,0 @@
-<?php
-
-namespace React\Tests\Nntp;
-
-class TestCase extends \PHPUnit_Framework_TestCase
-{
-    protected function expectCallableExactly($amount)
-    {
-        $mock = $this->createCallableMock();
-        $mock
-            ->expects($this->exactly($amount))
-            ->method('__invoke');
-
-        return $mock;
-    }
-
-    protected function expectCallableOnce()
-    {
-        $mock = $this->createCallableMock();
-        $mock
-            ->expects($this->once())
-            ->method('__invoke');
-
-        return $mock;
-    }
-
-    protected function expectCallableNever()
-    {
-        $mock = $this->createCallableMock();
-        $mock
-            ->expects($this->never())
-            ->method('__invoke');
-
-        return $mock;
-    }
-
-    protected function createCallableMock()
-    {
-        return $this->getMock('React\\Tests\\Nntp\\Stub\\CallableStub');
-    }
-}
