+++ tests/Rvdv/React/Tests/Nntp/Stub/CallableStub.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rvdv\React\Tests\Nntp\Stub;
+
+class CallableStub
+{
+    public function __invoke()
+    {
+    }
+}
