+++ tests/bootstrap.php
@@ -1,4 +1,4 @@
 <?php
 
 $loader = require __DIR__.'/../vendor/autoload.php';
-$loader->add('React\\Tests', __DIR__);
+$loader->add('Rvdv\\React\\Tests', __DIR__);
